Refactored to work scheduler and signature popup
